+++ src/ApiPlatform/State/StateProcessorTrait.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Dbp\Relay\CoreBundle\ApiPlatform\State;
 
 use ApiPlatform\Metadata\Delete;
+use ApiPlatform\Metadata\HttpOperation;
 use ApiPlatform\Metadata\Operation;
 use ApiPlatform\Metadata\Patch;
 use ApiPlatform\Metadata\Post;
@@ -18,7 +19,10 @@ trait StateProcessorTrait
     {
         $this->currentOperation = $operation;
         $this->currentUriVariables = $uriVariables;
-        $this->currentRequestMethod = $context['request']?->getMethod();
+
+        if ($operation instanceof HttpOperation) {
+            $this->currentRootRequestMethod = $operation->getMethod();
+        }
 
         if ($operation instanceof Post) {
             return $this->post($data, $context);