+++ src/ApiPlatform/State/StateProviderTrait.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Dbp\Relay\CoreBundle\ApiPlatform\State;
 
 use ApiPlatform\Metadata\CollectionOperationInterface;
+use ApiPlatform\Metadata\HttpOperation;
 use ApiPlatform\Metadata\Operation;
 use Dbp\Relay\CoreBundle\Exception\ApiError;
 use Dbp\Relay\CoreBundle\Rest\Query\Pagination\PartialPaginator;
@@ -22,7 +23,11 @@ trait StateProviderTrait
     {
         $this->currentOperation = $operation;
         $this->currentUriVariables = $uriVariables;
-        $this->currentRequestMethod = $context['request']?->getMethod();
+
+        $rootOperation = $context['root_operation'] ?? $operation;
+        if ($rootOperation instanceof HttpOperation) {
+            $this->currentRootRequestMethod = $rootOperation->getMethod();
+        }
 
         if ($operation instanceof CollectionOperationInterface) {
             return $this->getCollectionInternal($context);